pyproject.toml
[project]
name = "corona-account-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "djangorestframework-simplejwt>=5.5.1",
    "docker>=7.1.0",
    "gunicorn>=25.1.0",
    "pre-commit>=4.5.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "drf-spectacular>=0.29.0",
    "ruff>=0.15.1",
]
